Add shader, media content, and HLSL shader hooks
Provided shader is a white-based fog of war shader which treats the mouse
as the light source.
Media content was previously withheld via .gitignore, but it should not
grow too large in size due to the intended small size of this repo, so is
not terrible to commit. This is mainly to make showing the HLSL script
easier.